Add rudimentary variation support
There is a new method called back() in Game, which sets an internal
pointer to the current position back by one move.
doMove and getPosition work on this internal pointer instead of the
latest mainline move now.

Therefor it is theoretically possible to create a game with variations
now by first adding all the mainline moves and then going back to add
the variations.

However there isn't a forward() method yet, so variation support is
still somewhat incomplete.