Prevent change of owner_id on update
The previous commit introduced the problem that the owner_id of shared
resources would change when another user edited them.

This has been fixed by preventing any change of the owner_id via the
update() method. That means of course that even owners can't transfer
ownership of games and tags but I don't think this is a very common
scenario. Might still be added later though.