Only show accessible resources with index()
Users should only see resources accessible by them when they visit e.g.
/api/games

To accomplish that, a new concept called 'Filters' has been added to the
ExtendedRepository: Filters are applied on the result of a database
query, so they can access all of laravels collection methods instead of
relying on sql queries like Criteria do.

Known issues:
1. Filters aren't very robust, for example no type checking is done.
This should, however, be no problem as long as they are only used in
this project.
2. Accessing /api/games or /api/tags without being authenticated throws
an error. This will be fixed when public sharing of resources is
implemented.
3. /api/users is still completely accessible, meaning that all email
addresses are public. This will be fixed later by only showing summary
representations with public information. In the case of users only
showing the accessible ones is not planned.