Use Criteria instead of Filter for visibility
Filters have several problems: First of all the performance is much
worse, but there are also problems with pagination that would require
ugly hacks, so I've decided to use Criteria instead although that means
that the authorization code is now in two different places. Maybe I'll
try to use the SQL from the Criteria for regular authorization as well.