Ignore compiled and copied files in public
Since gulp must be run anyway after clonign this repo to copy the
libraries, the rest of the files should also be ignored by git.

This also has the advantage that there can never be an out-dated version
of the files in public/ (in case some one forgets compiling before
committing) causing weird bugs.