Fix #32
This is kind of a workaround but its the best solution I was able to
find for now.

The problem was that the height and width of the chess-board weren't set
yet when chessground cached them (I have no idea why that was only the
case for every tab after the first one).

To fix this, I simply cleared the boundary cache of chessground after
initializing the board. If the problem persists in some environments, we
might have to wrap that in a setTimeout() call to make sure the board
size is set.

If anyone has suggestions on how to fix this in a cleaner way, just
reopen #32 or submit a PR.