Support promotion in the UI
When a pawn reaches the 1st/8th rank, it is promoted to a queen by
default. At the same time, a small bar appears where the user can choose
another piece. If they do, that piece is used instead. But the bar can
also be ignored because the move is complete without any further action
(so e.g. exporting the game as a PGN would use the queen). If the next
move is made, the bar disappears again.